fullscreen/api/element-request-fullscreen-twice-manual.html

<!DOCTYPE html>
<title>Element#requestFullscreen() twice</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../trusted-click.js"></script>
<div id="log"></div>
<script>
async_test(t => {
  const div = document.querySelector("div");
  var rejected = false;

  document.onfullscreenchange = t.step_func(() => {
    assert_equals(document.fullscreenElement, div);
    // Done, but ensure that there's only one fullscreenchange event.
    document.onfullscreenchange = t.unreached_func("second fullscreenchange event");
    step_timeout(t.step_func_done());
  });
  document.onfullscreenerror = t.step_func(() => {
    assert_true(rejected, "second fullscreen fails");
    step_timeout(t.unreached_func('timer callback'));
    requestAnimationFrame(t.step_func_done());
  });

  trusted_click(t, () => {
    // Request fullscreen twice.
    div.requestFullscreen();
    assert_equals(document.fullscreenElement, null, "fullscreenElement after first requestFullscreen()");
    var p = div.requestFullscreen();
    if (p) {
      p.catch(t.step_func(() => rejected = true));
      Promise.resolve().then(t.step_func(() => assert_true(rejected)));
    }
  }, document.body);
});
</script>
